Move ServerResponse & ValuedEndpointOutput to server's model
