+++ doc/testing.md
@@ -89,7 +89,7 @@ following customised akka http options:
 import sttp.tapir.server.interceptor.exception.ExceptionContext
 import sttp.tapir.server.interceptor.CustomInterceptors
 import sttp.tapir.server.akkahttp.AkkaHttpServerOptions
-import sttp.tapir.server.ValuedEndpointOutput
+import sttp.tapir.server.model.ValuedEndpointOutput
 import sttp.model.StatusCode
 
 val customOptions: CustomInterceptors[Future, AkkaHttpServerOptions] =
