+++ metrics/opentelemetry-metrics/src/main/scala/sttp/tapir/metrics/opentelemetry/OpenTelemetryMetrics.scala
@@ -4,9 +4,10 @@ import io.opentelemetry.api.common.Attributes
 import io.opentelemetry.api.metrics.{DoubleHistogram, LongCounter, LongUpDownCounter, Meter}
 import sttp.tapir.AnyEndpoint
 import sttp.tapir.metrics.opentelemetry.OpenTelemetryMetrics._
-import sttp.tapir.model.{ServerRequest, ServerResponse}
+import sttp.tapir.model.ServerRequest
 import sttp.tapir.server.interceptor.metrics.MetricsRequestInterceptor
 import sttp.tapir.server.metric.{EndpointMetric, Metric, MetricLabels}
+import sttp.tapir.server.model.ServerResponse
 
 import java.time.{Duration, Instant}
