+++ server/core/src/main/scala/sttp/tapir/server/interceptor/RequestResult.scala
@@ -1,6 +1,6 @@
 package sttp.tapir.server.interceptor
 
-import sttp.tapir.model.ServerResponse
+import sttp.tapir.server.model.ServerResponse
 
 /** The result of processing a request: either a response, or a list of endpoint decoding failures. */
 sealed trait RequestResult[+B]
