+++ server/core/src/main/scala/sttp/tapir/server/interceptor/content/UnsupportedMediaTypeInterceptor.scala
@@ -2,12 +2,11 @@ package sttp.tapir.server.interceptor.content
 
 import sttp.model.{ContentTypeRange, StatusCode}
 import sttp.monad.MonadError
-import sttp.tapir._
+import sttp.tapir.{server, _}
 import sttp.tapir.internal._
-import sttp.tapir.model.ServerResponse
-import sttp.tapir.server.ValuedEndpointOutput
 import sttp.tapir.server.interceptor._
 import sttp.tapir.server.interpreter.BodyListener
+import sttp.tapir.server.model.{ServerResponse, ValuedEndpointOutput}
 
 /** If no body in the endpoint's outputs satisfies the constraints from the request's `Accept` header, returns an empty response with status
   * code 415, before any further processing (running the business logic) is done.
@@ -27,7 +26,7 @@ class UnsupportedMediaTypeInterceptor[F[_]] extends EndpointInterceptor[F] {
             val hasMatchingRepresentation = supportedMediaTypes.exists(mt => ranges.exists(mt.matches)) || supportedMediaTypes.isEmpty
 
             if (hasMatchingRepresentation) endpointHandler.onDecodeSuccess(ctx)
-            else responder(ctx.request, ValuedEndpointOutput(statusCode(StatusCode.UnsupportedMediaType), ()))
+            else responder(ctx.request, server.model.ValuedEndpointOutput(statusCode(StatusCode.UnsupportedMediaType), ()))
 
           case Left(_) =>
             // we're forgiving, if we can't parse the accepts header, we try to return any response
