+++ server/core/src/main/scala/sttp/tapir/server/interceptor/decodefailure/DecodeFailureInterceptor.scala
@@ -2,9 +2,9 @@ package sttp.tapir.server.interceptor.decodefailure
 
 import sttp.monad.MonadError
 import sttp.monad.syntax._
-import sttp.tapir.model.ServerResponse
 import sttp.tapir.server.interceptor._
 import sttp.tapir.server.interpreter.BodyListener
+import sttp.tapir.server.model.ServerResponse
 
 class DecodeFailureInterceptor[F[_]](handler: DecodeFailureHandler) extends EndpointInterceptor[F] {
   override def apply[B](responder: Responder[F, B], endpointHandler: EndpointHandler[F, B]): EndpointHandler[F, B] =
